Implement `TemplateStringMixin`
