+++ java-runtime/src/main/java/io/spine/validate/diags/ViolationText.java
@@ -36,7 +36,6 @@ import java.util.Collection;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.string.Diags.backtick;
-import static io.spine.validate.TemplateStringExtsKt.format;
 import static java.lang.System.lineSeparator;
 import static java.util.stream.Collectors.joining;
 
@@ -80,7 +79,7 @@ public final class ViolationText {
         var fieldPath = path.getFieldNameCount() == 0
                         ? ""
                         : Field.withPath(path).toString();
-        var formattedMessage = format(violation.getMessage());
+        var formattedMessage = violation.getMessage().format();
         var result = new StringBuilder();
         appendPrefix(result, typeName);
         appendPrefix(result, fieldPath);
