+++ java-runtime/src/main/kotlin/io/spine/validate/TemplateStringExts.kt
@@ -36,53 +36,6 @@ import io.spine.validate.RuntimeErrorPlaceholder.PARENT_TYPE
 import io.spine.validate.RuntimeErrorPlaceholder.REGEX_PATTERN
 
 /**
- * Returns a template string with all placeholders substituted with
- * their actual values.
- *
- * For example, for a template string with the following values:
- *
- * ```
- * with_placeholders = "My dog's name is ${dog.name}."
- * placeholder_value = { "dog.name": "Fido" }
- * ```
- *
- * This method will return "My dog's name is Fido."
- */
-public fun TemplateString.format(): String {
-    checkPlaceholdersHasValue(withPlaceholders, placeholderValueMap) {
-        "Cannot format the given `TemplateString`: `$withPlaceholders`. " +
-                "Missing value for the following placeholders: `$it`."
-    }
-    return formatUnsafe()
-}
-
-/**
- * Returns a template string with all placeholders substituted with
- * their actual values, without validating that all placeholders have
- * corresponding values.
- *
- * This method does not check whether every placeholder in the template has a matching value
- * in the placeholder map. Any placeholders without a corresponding value will remain
- * unchanged in the resulting string.
- *
- * For example, for a template string with the following values:
- *
- * ```
- * withPlaceholders = "My dog's name is ${dog.name} and its breed is ${dog.breed}."
- * placeholderValue = { "dog.name": "Fido" }
- * ```
- *
- * This method will return "My dog's name is Fido and its breed is ${dog.breed}.".
- */
-public fun TemplateString.formatUnsafe(): String {
-    var result = withPlaceholders
-    for ((key, value) in placeholderValueMap) {
-        result = result.replace("\${$key}", value)
-    }
-    return result
-}
-
-/**
  * Makes sure that each placeholder within the [template] string has a value
  * in [placeholders] list.
  *
