+++ java-runtime/src/main/proto/spine/validate/error_message.proto
@@ -55,6 +55,8 @@ option java_package = "io.spine.validate";
 //
 message TemplateString {
 
+    option (is).java_type = "io.spine.validate.TemplateStringMixin";
+
     // The template string that may contain one or more placeholders.
     string with_placeholders = 1;
